Implementation of DrawTriangleTrue and DrawTriangleFast
DrawTriangleTrue uses an exhaustive scan and matrix work to determine
which pixels are in a triangle or not.  Accurate and true, but very slow
at the moment.

DrawTriangleFast is fairly close to accuracy, but there are odd edge
cases that can occur and error.  But...  It uses a variant of the
Bresenham Algorithm to find the two X points along the X-axis for each Y
and then memcpy a pixel buffer to the space between.  Close to 20+ times
faster at the moment.