General Cleanup and Optimization of OpenGL Drawing
Simplified the OpenGL drawing routine and moved commands that were needed within the drawing loop.  Only one texture to worry about, so...

General Cleanup to naming scheme to make it cleaner.  Created functions to handle refresh locking instead of directly accessing the mutex.  This allows the header to be cleaner.