Experimental Tweaks and Oddball Fixes
Some experimental tweaks have been done to improve performance,
functionality, and flow.  A few are a bit hacky, but not too offensive.
Still much clean up to be done.

A few fixes here and there, and some general capabilities added.  Added
a Back Buffer for memory scratch spaces that is the same size as the
main video memory.  And added a row buffer to give a scratch space for
pixel row memcpy operations.  That way you don't have to waste time on
malloc and free operations.