Experimental OpenGL Persistent Buffer Mapping Support
Using Persistent Buffer Mapping to a Pixel Buffer Object, speed issues
may have been resolved.  Needs to be verified.  If so, this will be the
preferred video rendering method and the old method will be used as a
fall back in case the right version of the OpenGL is not supported.